start-stop: Add support for password input
If a oneshot needs to ask for user input (password, to e.g. luksOpen)
they can simply write "< " followed by the message asking the password
to the fd 3, and read on their stdin (fd 0).

anopa will handle the prompt properly, and write the user input on the
process' stdin. Simple as that.