stop: Process "needs" in reverse order
We used to only stop whatever was specified, processing dependencies
("needs") only as ordering directives (i.e. as "after"), which obviously
isn't right.

Now they're correctly processed "in reverse," i.e. if A needs B then we'll
set it as B needs A, i.e. to stop B we also need to stop A (and the
matching "after" will be there to order stopping A before B).