enable: Add --upgrade to upgrade servicedirs
repodir & servicedirs must exists; Then we empty the servicedir (save for
status.anopa, down, supervise/ and event/) and do as usual.
We also don't create file "down" or symlink in scandir for longruns.

For auto-added services (from needs/wants) we only check they exists (i.e. are
enabled): if so we do nothing, else we do as usual/as if w/out --upgrade

Also --set-{crash,finish} are ignored when --upgrade is used.