Tweak how --double-output is handled
Remove aa_init_output() to instead do the istty() calls when needed, and
simply have a aa_output_setboth() to enable(/disable) double output mode.

Makes things a bit simpler, and will allow to have it accounted for right
away (i.e. when option is processed) which will be useful when we're
using our own functions for strerr_* errors (to support double output).