Support --double-output about everywhere
aa-sync does not, but it doesn't ever output anything so it's ok.

This is useful for even warnings/errors (from aa_strerr_*) when stderr
is redirected to a file. Note that parsing order matters, so put your -D
first.

Note: aa-setready had its -D switch to -N for --unready, better
consistency that way. (Also with the latest s6 there's a new event 'D'
so ours will be renamed to 'N' as well, or just removed.)