start/stop: Add --verbose (LOADFAIL->AUTOLOAD)
Rename AA_LOADFAIL_* to AA_AUTOLOAD_* and aa_loadfail_cb to aa_autoload_cb
and trigger the callback whenever a service is auto-added, not only on
error (The err argument will be zero on success).

Also trigger it on MODE_STOP, though there as usual things are "reversed":
si is of the service that is needed, and name of the main service.
IOW if foo needs bar, in MODE_STOP it will be bar's si and name="foo"
since (to stop) bar needs (to first stop) foo

This allows to add --verbose in both aa-{start,stop}