stop: Fix unable to stop service that keeps crashing
That is, when a service keeps crashing/stopping, and being restarted by its
supervise, we would see it as down and therefore not (allow to) stop it.

Now consider "to be restarted" to be up when stopping, to solve the issue. The
command will be sent to supervise (and the down file created), but we won't wait
for a 'd' event and announce it as "Not up".