Fix possible memory corruption
openreadfileclose() means to automatically enlarge the given stralloc
according to the passed limit *if needed*, i.e. it will not necesarilly
do so; It might not enlarge it, or only to fit the data read from the
file.

So assuming that svst->sa had been enlarged up to the possible limit was
wrong, and could result in writing to svst->sa.s[svst->sa.len] as out of
bounds write, leading to memory corruption.

Thanks to John O'M. for reporting & tracking this; Closes #5