enable: Add --alarm-s6
When adding new servicedirs to an active repodir (i.e. its scandir has a
running s6-svscan), for new longruns to be properly set/usable we need to
"alarm" s6-svscan, so it rescans the scandir.

This had to be done manually via `s6-svscanctl -a $REPODIR/.scandir` but
can now be done automatically from aa-enable.

Closes #6