start/stop: Fix possible warning "no service for id"
A warning "unable to find longrun service for id#x" could be shown at
times. This was because we first subscribe to a service, then check its
current status and if what we want, unsubscribe & process the service
as already up/not up.

However, for one-time subscription if the event happens to have been
triggered right in between, then we're already unsubscribed and the
event for that id already queued up.

So, we now check that unsubscribing was successful, and if not because
the id isn't of a subscription anymore then we'll properly treat the
event later as usual.