Minor fix..
..or should I say, incomplete.

Really this whole flush/noflush thing is a mess, and should be addressed
properly. We need to just have "regular" and flush functions, the later
to force flushing when needed.

But at least for now, this should fix issue when the buffer gets full
and we're noflushing, resulting in a messy/truncated output.