Fix output possibly being truncated..
..because we used _noflush() functions. There's really no need to, so
instead we have "regular" functions (that will be noflush unless the
buffer gets full, then it's auto-flushed) and flush functions to force
flushing when needed.

Much better, in a few cases (e.g. aa-status -La) that will be less
write calls, and we avoid possibly having truncated output (because we
used noflush & filled the buffer)