start/stop: Add support for retries
Number of retries can now be specified via regular file "retries" in
servicedir. It works for:

- oneshots: if start fails, we re-try up to RETRIES times before
  processing the service as failed;
- longruns with readiness support: if we get event 'd' we "allow" it/keep
  waiting (for 'U') up to RETRIES times before processing the service as
  failed.
  This is only for this transaction's purpose, i.e. the supervised state
  isn't affected, and s6-supervise will keep re-starting the service.

Note that until now, we kept "allowing" events 'd' and waiting forever
for event 'U' -- now by default on the first 'd' we stop, and there's no
way to wait forever, a specific number of retries must be specified.