Fix possible segfault
If a service were to end and its fd_progress was closed whilst a password
input was in progress, we would close_fd_for() it before calling
end_si_password(), which would then try to work with a reset struct
progress.

So make sure when we close the fd_progress for the si_password, we call
end_si_password() to properly end things.